saving & loading voxel maps

+++ engine/voxel/globals.go
@@ -1,7 +1,7 @@
 package voxel
 
 const (
-	EMPTY              int32 = 0
+	EMPTY              byte = 0
 	CHUNK_SIZE         int32 = 32
 	CHUNK_SIZE_SQUARED int32 = CHUNK_SIZE * CHUNK_SIZE
 	CHUNK_SIZE_CUBED   int32 = CHUNK_SIZE * CHUNK_SIZE * CHUNK_SIZE
